requirements.txt
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.7.1
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httptools==0.6.1
idna==3.6
msgpack==1.0.8
oauthlib==3.2.2
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.2
requests==2.31.0
requests-oauthlib==1.3.1
sniffio==1.3.1
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
